sassのベース作成、headerを別コンポーネントに分ける
