+++ _src/scss/_settings.scss
@@ -1,0 +1,66 @@
+@charset "utf-8";
+// /*=============================================================
+// 	Settings
+// =============================================================*/
+$COLOR_MAP: (
+	white: #fff,
+	red: #d9333f,
+	gray-lv1: #333,
+	gray-lv2: #999,
+	gray-lv3: #ccc,
+	gray-lv4: #e1e1e1,
+	key1: #7a4171,
+	key2: #007bbb
+);
+
+// COLOR THEME
+$C_WHITE: map-get($COLOR_MAP, white);
+$C_KEY1: map-get($COLOR_MAP, key1);
+$C_KEY1_LIGHT: lighten(map-get($COLOR_MAP, key1), 55%);
+$C_KEY2: map-get($COLOR_MAP, key2);
+$C_GRAY_LIGHT: map-get($COLOR_MAP, gray-lv3);
+$C_GRAY_BG: map-get($COLOR_MAP, gray-lv4);
+$C_TEXT: map-get($COLOR_MAP, gray-lv1);
+$C_LINK: map-get($COLOR_MAP, key2);
+
+// FONT FAMILIY
+$FONT_BASE: 'Hiragino Kaku Gothic ProN', YuGothic, 'Yu Gothic Medium', '游ゴシック', 'Yu Gothic', Meiryo, sans-serif;
+$FONT_SERIF: 'Yu Mincho Demibold', 'YuMincho', 'Hiragino Mincho ProN W3', 'ヒラギノ明朝 ProN W3', 'Hiragino Mincho ProN', 'HG明朝E', serif;
+$FONT_SERIF_BOLD: 'Yu Mincho Demibold', 'YuMincho Demibold', 'Hiragino Mincho ProN W6', 'ヒラギノ明朝 ProN W6', 'Hiragino Mincho ProN', 'HG明朝E', serif;
+
+// media
+$BREAKE_POINT_PC: 768px;
+$BREAKE_POINT_MOBILE_LANDSCAPE: 440px;
+$BREAKE_POINT_MOBILE_PORTRAIT: 375px;
+@mixin mqPc {
+  @media only screen and (min-width: $BREAKE_POINT_PC) {
+    @content;
+  }
+}
+@mixin mqMobile {
+  @media only screen and (max-width: $BREAKE_POINT_PC - 1) {
+    @content;
+  }
+}
+@mixin mqMobilePortrait {
+  @media only screen and (max-width: $BREAKE_POINT_MOBILE_LANDSCAPE - 1) {
+    @content;
+  }
+}
+@mixin mqMobileSmall {
+  @media only screen and (max-width: $BREAKE_POINT_MOBILE_PORTRAIT - 1) {
+    @content;
+  }
+}
+
+// width
+$WIDTH_CONTENTS: 1072px;
+$CONTENTS_PADDING_SP: 16px;
+$CONTENTS_PADDING_SP_SMALL: 10px;
+
+// animation
+$TRANSITION_HOVER_BASE: 0.3s ease-out;
+$OV_OPACITY: 0.7;
+
+// border
+$BORDER_RADIUS_BASE: 5px;
