+++ index.html
@@ -9,7 +9,6 @@
   <link rel="stylesheet" href="css/common.css">
 </head>
 <body>
-  <h1 class="bg-secondary text-white h4 p-3">ミニ投稿アプリ（仮）</h1>
   <div class="container">
     <div id="app"></div>
   </div>
