+++ public/css/common.css
@@ -1,3 +1,148 @@
+@charset "UTF-8";
+* {
+  padding: 0;
+  margin: 0;
+}
+
+section, article, aside, hgroup, header, footer, nav, figure, figcaption, summary, main {
+  display: block;
+  padding: 0;
+  margin: 0;
+}
+
+html {
+  -webkit-text-size-adjust: 100%;
+}
+
 body {
-  background-color: #ccc;
+  color: #333;
+  font: normal normal 16px/1 Hiragino Kaku Gothic ProN, YuGothic, Yu Gothic Medium, 游ゴシック, Yu Gothic, Meiryo, sans-serif;
+  font-feature-settings: 'pwid';
+  letter-spacing: 0.04em;
+  background: #fff;
+  text-align: left;
+}
+
+ol, ul, dl, li, dt, dd {
+  list-style: none;
+}
+
+img {
+  vertical-align: top;
+}
+
+a, img {
+  border: none;
+}
+
+h1, h2, h3, h4, h5, h6 {
+  font-size: 100%;
+}
+
+sup, sub {
+  height: 0;
+  line-height: 1;
+  vertical-align: baseline;
+  margin: 0 1px;
+  position: relative;
+}
+
+sup {
+  bottom: 0.7ex;
+}
+
+sub {
+  top: 0.1ex;
+}
+
+a {
+  color: #007bbb;
+  text-decoration: none;
+  transition: 0.3s ease-out;
+  transition-property: color, background, opacity;
+}
+
+a:hover {
+  text-decoration: underline;
+}
+
+img {
+  box-sizing: border-box;
+  max-width: 100%;
+  height: auto;
+}
+
+img.-border {
+  border: 1px solid #ccc;
+}
+
+@media only screen and (max-width: 767px) {
+  .pc:not(.sp) {
+    display: none !important;
+  }
+}
+
+@media only screen and (min-width: 768px) {
+  .sp:not(.pc) {
+    display: none !important;
+  }
+}
+
+@media only screen and (min-width: 768px) {
+  .spP:not(.pc) {
+    display: none !important;
+  }
+}
+
+@media only screen and (min-width: 375px) and (max-width: 767px) {
+  .spP {
+    display: none !important;
+  }
+}
+
+#container {
+  min-width: 320px;
+}
+
+.msgOver textarea,
+.msgOver textarea:focus {
+  color: #d33;
+  background-color: #fafadf;
+}
+
+.msgOver .msg-counter > span {
+  color: #d33;
+  font-weight: bold;
+}
+
+.btn-like {
+  display: inline-block;
+  width: 20px;
+  height: 20px;
+  line-height: 1;
+  cursor: pointer;
+}
+
+.btn-like path {
+  fill: #d33;
+  stroke-width: 50;
+  stroke: #d33;
+  fill-opacity: 0;
+  transform: scale(0.9, 0.9);
+  transform-origin: center center;
+  transition: .3s;
+}
+
+.btn-like:hover path {
+  fill-opacity: .2;
+}
+
+.btn-like.on path {
+  fill-opacity: 1;
+}
+
+.like-cnt {
+  line-height: 20px;
+  display: inline-block;
+  padding-left: 5px;
 }
