+++ src/App.vue
@@ -1,14 +1,20 @@
 <template>
-  <router-view />
+  <Header />
+  <main>
+    <div class="contents">
+      <router-view />
+    </div>
+  </main>
 </template>
 
 <script>
+import Header from './components/Header.vue'
 import Home from './components/Home.vue'
 
 export default {
   name: 'App',
   components: {
-    Home
+    Header, Home
   }
 }
 </script>
