+++ src/main.js
@@ -1,6 +1,5 @@
 import { createApp } from 'vue'
 import App from './App.vue'
-import './index.css'
 import { router } from './router'
 import { store } from './store'
